feat-#5: rename repository to repositories
