+++ src/item/repositories/item-write.repository.ts
@@ -1,6 +1,6 @@
 import { EntityRepository, Repository } from "typeorm";
 
-import { ItemEntity } from "../entity";
+import { ItemEntity } from "../entities";
 
 @EntityRepository(ItemEntity)
 export class ItemWriteRepository extends Repository<ItemEntity> {}
