--- src/item/repository/item.repository.ts
@@ -1,14 +1,0 @@
-import { Injectable } from '@nestjs/common';
-
-import { Item } from '../models';
-
-@Injectable()
-export class ItemRepository {
-  deleteItem(id: number) {
-    const item = new Item(id);
-
-    item.deleteItem();
-
-    return item;
-  }
-}
